Improved Missing Field Messages
Previously if a field was unpublished you'd get a vague ArgumentNullException from Sitecore upon accessing a Synthesis item. This rectifies this by passing the requisite data through to display a decent error message if a field that does not exist in Sitecore is accessed.

This will also throw a MissingTemplateFieldException, so it can be specifically caught if need be.