Allow Force Rebuilding
With previous releases of Synthesis, if you made configuration file changes (eg updated namespace or excluded templates) the handler page would not detect them as a desynchronized model and refuse to regenerate.

Synchronization is still exclusively a field and template level affair, however the option to force a regeneration even if a 'synchronized' model is available has been added.