Add ModelOutputBasePath
This setting allows you to specify a global base path which all configurations' model paths will use as a base path.
A common usage of this setting would be for out of webroot development where you want to specify the source solution path
as a base and the relative path within the solution as the model path.

It it also possible to directly specify the base paths, however a setting can accept sc.variable values (e.g. $(sourcePath) for Habitat, $(dataFolder))
whereas the model path cannot (Sitecore does not expand variable values in XML content, only attributes).

If this value is blank, it is ignored. Otherwise it is prefixed to the ModelOutputPath (e.g. this = c:\foo\, ModelOutputPath = src\Bar\Model.cs, resulting path = c:\foo\src\Bar\Model.cs)

Note that the path is directly prepended so you will need a trailing slash here, or a leading slash in the model path, to ensure a validly formatted path