Bugfix for solr field name translation
While trying to use a LINQ query with GetSynthesisQueryable I noticed that the field name translation for solr was not working as intended:
A query.Where(x => x.Title.RawValue) was translated to the following solr query: title_t.rawvalue_t (RawValue was translated too). I solved this by wrapping the SynthesisSolrFieldNameTranslator in the regular SynthesisFieldNameTranslator, so the "normal" field name translation occurs also when using solr. I hope, this makes sense, but so far I had no issues with it.